requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
ffmpeg==1.4
fpdf==1.7.2
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
multidict==6.0.5
mutagen==1.47.0
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
pipdeptree==2.13.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
PyNaCl==1.5.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
validators==0.22.0
vboxapi==1.0
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
youtube-search==2.1.2
yt-dlp==2023.12.30
